tox -e test: allow passing arguments.
For example: `tox -e test -- -D` to get a debug prompt after a test failure.
Note that you need `--` to separate the options that you pass to tox from the options that you pass to the command in the tox environment.
See also https://tox.wiki/en/latest/config.html#substitutions-for-positional-arguments-in-commands