feat(pre-commit): check python versions
Add another linter on `pre-commit`: https://pypi.org/project/check-python-versions/

It checks that all python versions specific across the repository are
consistent.

For now, we only check it on `setup.py` and `pyproject.toml` as we are
not adding on `tox.ini` all python versions, as we already have Jenkins
for that, at least for now.