feat: avoid global variables on GHA
Avoid depending on `vars.*` variables defined at GitHub repository or
organization level to run the test job.

We need a way to define which python and OS combinations each repository
wants to use, but not defined in a global way, that by default, is
broken for everyone that is not in Plone GitHub organization.