fix(gha/gitlab): handle only one python version
If a custom `test_matrix` is specified with a single python version for
a given Plone version, do not duplicate the GHA/GitLab jobs created.

`plone.meta` generates GHA/GitLab jobs only for the python lower and upper
bounds, but if there is only one Python version, it was generating the
same GHA/GitLab job twice, which is reported as an error in GitHub and
GitLab.